Submit captcha_key in hidden_field_tag
simple_captcha submits the captcha key (for the DB entry) in the controller based method in the session only. In the model based method, it's submitted in a hidden field. This patch submits the captcha key in a hidden field in the controller based method, too, so it works without cookies.